add ajout de commentaire

+++ src/main/java/com/jee/projet/Controller/ProgramController.java
@@ -30,7 +30,7 @@ public class ProgramController {
         User user = userService.getUserById(21);
         Program p= progamService.getById(id);
         model.addAttribute("programme",p);
-        float i=progamService.getMoyennne(p,user);
+        double i=progamService.getMoyennne(p,user);
         model.addAttribute("moyenne",i);
         return "programme";
     }
